# make encryption/decryption faster with big files